xen/kbdif: Add frontend's virtual device options
It is now not possible to control if and which virtual devices
are created by the frontend, e.g. keyboard and pointer devices
are always created and multi-touch device is created if the
backend advertises multi-touch support. In some cases this
behavior is not desirable and better control over the frontend's
configuration is required.

Add new frontend options, so it is possible to individually
control set of exposed virtual devices for each guest OS:
 - set opt-no-pointer if no pointer device needs to be created
 - set opt-no-keyboard if no keyboard device needs to be created
 - set opt-no-multi-touch if no multi-touch device needs to be created

Keep old behavior by default.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>