vpci: add tear down functions
Tear down functions are not mandatory. Note that this patch just
implements the framework, but doesn't implement any tear down function
yet.

The list of vpci init and teardown functions is also moved outside of
the init section. This is in preparation for SR-IOV support, which
requires the ability to add and remove pci devices (and vpci handlers)
after boot.

No functional change intended.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>