vpci/sriov: add support for SR-IOV capability
So that a PCI device that supports SR-IOV (PF) can enable the capability
and use the virtual functions.

This code is expected to only be used by privileged domains,
unprivileged domains should not get access to the SR-IOV capability.

The current code detects enabling of the virtual functions feature and
automatically adds the VFs to the domain. It also detects enabling of
memory space and maps the VFs BARs into the domain p2m. Disabling of
the VF enable bit removes the devices and the BAR memory map from the
domain p2m.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>