HACK: arm/p2m: do not invalidate p2m root if it is shared with IOMMU
In case if the p2m table is shared to IOMMU, invalidating it turns IOMMU to
translation faults that could be not repaired.

Fixed patch check for the corresponded condition and has a comment for one
introduced p2m_invalidate_root() call, but miss them for another. So put them
in place.

Despite said above, this should not be aplied as JG said. Driver should be
modified to not use Set/Way instead.

Fixes: 2148a12 ("xen/arm: Track page accessed between batch of Set/Way operations")
Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>