arm: Implement has_vpci macro
Currently (as of ARM PCI series) has_vpci macro is always set to
true and actual decisions are made on if has_vpsi is true and
if the domain is a hardware one or not.
This doesn't work for driver domains.

Assume that if "vpci" configuration option is present in domain's
xl config file then vpci is enabled for the domain and has_vpci
must return true, otherwise, return false. This allows setting
vpsi only for those domains which want that.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>