[hwdom DomD] xen/arm: allow binding PCI bridges with "xen,passthrough" set
While running Xen with pci-passthrough=on boot argument Xen tries
to instantiate PCI devices and bind them to their drivers. One of
the checks while matching compatible devices is a check if that
device is a part of Domain-0 (e.g. device is enabled in Xen device
tree and it has no "xen,passthrough" property set).

PCI host bridge can live in a driver domain, so Domain-0 won't own
it ("xen,passthrough" will be set in that case). For this specific case
of PCI still allow Xen binding the driver even if "xen,passthrough" is set
(for devices with device_type = "pci" in the corresponding device tree node).

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>