xen/gnttab: fixed crash when grant mappings is used for DMA requests
dev_bus_addr returned by the hypercall is MFN when device is not
protected by IOMMU. Guest can do the DMA requests successfully.

If device is protected by IOMMU, the request will fail as there is no
mapping for grant table in IOMMU page tables.

If page tables is not shared between IOMMU and CPU, we can add the
mapping for the grant table in IOMMU page tables by calling
iommu_map/iommu_unmap functions, but if page table is shared
the solution is to return IPA address.

Signed-off-by: Rahul Singh <rahul.singh@arm.com>