iommu/ipmmu-vmsa: Use refcount for the micro-TLBs
Reference-count the micro-TLBs as several bus masters can be
connected to the same micro-TLB. This wasn't an issue so far,
since the platform devices (this driver deals with) get
assigned/deassigned together during domain creation/destruction.
But, to support PCI devices (which are hot-pluggable) we will
need to take care of.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>