iommu/arm: Introduce iommu_add_pci_device API
The main puprose of this patch is to add a way to register PCI device
(which is behind the IOMMU) using the generic PCI-IOMMU DT bindings [1]
before assigning that device to a domain.

This behaves in almost the same way as existing iommu_add_dt_device API,
the difference is in devices to handle and DT bindings to use.

The function of_map_id to translate an ID through a downstream mapping
(which is also suitable for mapping Requester ID) was borrowed from Linux
and slightly updated according to the Xen code base.

TODO: Consider whether we need an alternative to pci_for_each_dma_alias.

[1] https://www.kernel.org/doc/Documentation/devicetree/bindings/pci/pci-iommu.txt

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>

---
While rebasing this patch on the integration branch I noticed that
the code to manage "iommu-map/iommu-map-mask" is already present
in iommu_helpers.c, although it is not connected to add_device and
with no real usage in current code base. It was added by
"xen/arm : Add PCI device support for SMMU".
---