vpci/header: Implement SR-IOV BAR handling
Emulate guest BAR register values based on PF BAR values for VFs.
This allows creating a guest view of the normal BAR registers and emulates
the size and properties as it is done during PCI device enumeration by
the guest.

Note, that VFs ROM BAR is read-only and is all zeros, but VF may provide
access to the PFs ROM via emulation and is not implemented.

SR-IOV has PCI_SRIOV_NUM_BARS number of VF BARs, but for
the p2m map/unmap machinery code be simpler and stay unified
we provide more BARs here then needed: this must stay in sync with the number
of BARs defined in the struct vpci_header and is enforced by
BUILD_BUG_ON compile time check.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>