vpci/header: use pdev's domain for deferred mapping
When we defer map/unmap operations for a guest domain it is possible
that vCPU's domain is different from pdev's one. If this is the case
then a trap is observed for the guest as the MMIO range is not mapped
for it:

(d1) r8169 0000:00:00.0: enabling device (0400 -> 0402)
(XEN) map [71700, 71700] -> 0x23004 for d0
(XEN) traps.c:2014:d1v0 HSR=0x00000093950006 pc=0xffffffc010782e20 gva=0xffffffc0112e3040 gpa=0x00000023004040

Fix this by referencing the proper domain while doing the deferred
map/unmap operations.

Fixes: 9c244fdef7e7 ("vpci: add header handlers")

Signed-off-by: Rahul Singh <rahul.singh@arm.com>
Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
---
Since v3:
- re-introduced in v4 after being dropped in v2: added full description
  of the bug and its fix