iommu/ipmmu-vmsa: Clear USE_SECGRP bit in USE_SECGRP register
We need to make sure that USE_SECGRP bit is not set (the initial
value is unknown) as DMAC doesn't support USE_SECGRP function.
It is still a question whether this really needs to be done in Xen,
rather in firmware, but better to be on the safe side. After all,
if firmware already takes care of it, nothing bad will happen.

Based on the following commits from the Renesas BSP:
8fba83d iommu/ipmmu-vmsa: Set IPMMU bit IMSCTLR_USE_SECGRP to 0
a8d93bc iommu/ipmmu-vmsa: Update IMSCTLR register offset address
for R-Car S4
located at:
https://github.com/renesas-rcar/linux-bsp/tree/v5.10.41/rcar-5.1.3.rc5

Please note, this is not a direct port of above commits because
on Xen ipmmu_domain_init_context() only setups Root IPMMU registers
for the context in question (Cache IPMMU registers are not touched
there). Moreover, this is not a MMU "context" register at all,
so there is no point in keeping the clearing there. Instead, perform
the clearing at once during initialization.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>