[NEW UPSTREAM] xen/arm: add pci-domain for disabled devices
If a PCI host bridge device is present in the device tree, but is
disabled, then its PCI host bridge driver was not instantiated.
This results in the failure of the pci_get_host_bridge_segment()
and the following panic during Xen start:

(XEN) Device tree generation failed (-22).
(XEN)
(XEN) ****************************************
(XEN) Panic on CPU 0:
(XEN) Could not set up DOM0 guest OS
(XEN) ****************************************

Fix this by adding "linux,pci-domain" property for all device tree nodes
which have "pci" device type, so we know which segments will be used by
the guest for which bridges.

Fixes: 4cfab4425d39 ("xen/arm: Add linux,pci-domain property for hwdom if not available.")

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>

---
This patch aims for 4.16 release.
Benefits:
Fix the bug for the platforms that have multiple definitions for PCI
host bridges yet some of them are disabled and not intended to be used,
development.
Risks:
No risk for the platforms that do not define PCI host bridges with
status = "disabled" (current behavior).