[DOMD][HACK] initialize vpci right before startng the domain
In vpci_domain_init we rely on used_by property for the PCI host bridges
to be set in order to detect if this is a PCI hardware domain.
The start sequence is that vpci_domain_init is called first and then
the toolstack assigns devices with XEN_DOMCTL_assign_device and used_by
is properly set. This leads to the wrong assumption on PCI hardware
domain and results in the fact that MMIO handlers for the DomD are set
as for other guest domain, e.g. with virtual ECAM host bridge.

Work this around by assigning MMIO handlers late right before jumping to
the guest.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>