Revert "[DOMD] ipmmu: Honor "xen,force-assign-without-iommu" property"
Not needed anymore with PCIE-IPMMU OSID support in place. We decided to
provide OSID0 (which is default) for DomD exclusively (athough it
doesn't seem required for 0000:00:00.0 device, it could be used for any
devices in DomD without manual BDF-OSID assignment). So we can restore
"iommus" property for the host bridge node which allows us to remove
the hack. It's fine for now but can be reconsidered later on if we suffer
from the lack of available OSIDs.

This reverts commit 3e49e97de3838c8c69896309b0610fa891a5eef2.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>