pci: convert pcidevs lock into read/write lock
Currently pcidevs lock is a global spinlock which is fine for
the existing use cases. It is used to both protect pdev instances
themselves from being removed while in use and to make sure the update
of the relevant pdev properties is synchronized.

Moving towards vPCI is used for guests this becomes problematic in terms
of lock contention. For example, during vpci_{read|write} the access to
pdev must be protected to prevent pdev disappearing under the feet and
this needs to be done with the help of pcidevs_{lock|unlock}.
On the other hand it is highly undesirable to lock all other pdev accesses
which only use pdevs in read mode, for example those which do not remove
or add pdevs.

For the above reasons change pcidevs spinlock to a read/write lock which
will help preventing locking contentions between pdev readers and writers.
Not to mix pcidevs lock use through the code remove
pcidevs_{lock|unlock} pair and introduce dedicated reader and writer
helpers.

Modify the existing code to use read and write locks where appropriate:
- most of the code will use pcidevs_write_{lock|unlock}
- distinct uses of pci_get_real_pdev and pci_get_pdev are converted
  to pcidevs_read_{lock|unlock}

N.B.: As of the XSA-114 fix we don't allow reader recursion, so readers
must not recursively read lock on the same CPU. This was now
guaranteed as pcidevs lock was not a recursive lock.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>